Migrate test/loaders/*.js test files to Vitest
